// src/shared/dataprovider/api/model/createUserRequestDTO.ts
/**
 * Dynasty Cash In Gateway
 * Dynasty Cash In API list
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AgentStatus } from './agentStatus';
import { AgentTypeEnum } from './agentTypeEnum';
import { UserRoleEnum } from './userRoleEnum';

export interface CreateUserRequestDTO { 
    Username?: string;
    Password?: string;
    Email?: string;
    Name?: string;
    RoleId?: number;
    ClientId?: number;
    AgentType?: AgentTypeEnum;
    Status?: AgentStatus;
    ParentAgentId?: number;
    IsAgent?: boolean;
    BankWithdrawalFee?: number;
    USDTWithdrawalFee?: number;
    CashPickUpWithdrawalFee?: number;
}